Be explicit about the parameter type of the array version of make_unique.

+++ win32/spinning-wings-gl2/Program.cpp
@@ -21,7 +21,7 @@ namespace silnith::wings::gl2
         GLint logSize{ 0 };
         glGetProgramiv(id, GL_INFO_LOG_LENGTH, &logSize);
         if (logSize > 0) {
-            std::unique_ptr<GLchar[]> log{ std::make_unique<GLchar[]>(logSize) };
+            std::unique_ptr<GLchar[]> log{ std::make_unique<GLchar[]>(static_cast<std::size_t>(logSize)) };
             glGetProgramInfoLog(id, static_cast<GLsizei>(logSize), nullptr, log.get());
             linkLog = { log.get() };
         }
